requirements.txt
pyshorteners==1.0.1
python-telegram-bot==13.11
